Break commands into full code blocks for easier cp
As I was reading through the docs I noticed some of the commands to get up
and running were on new lines and some where inline with the instructions.
This is an attempt to conform them all to be on new lines in their own
code blocks. This makes it easier to copy and paste the commands as
someone walks through the setup.